create folder src/styled, take out styled components from jsx components, put them into src/styled. created variable for global colors, threw props to components
