+++ src/components/FlexUl.jsx
@@ -1,21 +1,8 @@
 import React from 'react'
 
+import { StyledFlexUl } from "../styled/components/FlexUl"
 
-import styled, { css } from "styled-components";
 
-const StyledFlexUl = styled("ul")`
-  display: flex;
-  flex-wrap:wrap;
-  flex-direction: ${(props) => props.direction || "row"};
-  align-items: ${(props) => props.align || "stretch"};
-  justify-content: ${(props) => props.justify || "stretch"};
-  margin: ${({ margin }) => margin || "0"};
-  ${(props) =>
-    props.underline &&
-    css`
-      border-bottom: 1px solid #fff;
-    `}
-`;
 const FlexUl=(props)=> <StyledFlexUl {...props}/>
 
 export default FlexUl
