+++ src/components/Title.jsx
@@ -1,15 +1,8 @@
 import React from 'react'
-import styled from 'styled-components'
+
+import { StyledTitle } from "../styled/components/Title";
 
 
+const Title = (props) => <StyledTitle {...props} />;
 
-const Name = styled("h1")`
-  color: #fff;
-  text-align:center;
-  font-weight: 500;
-  
-  
-`;
-const Title=(props) =><Name {...props}/>
-
-export default Title
+export default Title;
