+++ src/styled/components/Flex.js
@@ -1,0 +1,16 @@
+import styled, {
+  css
+} from 'styled-components';
+
+
+export const StyledFlex = styled('div')
+`
+  display: flex;
+  flex-direction: ${(props) => props.direction || 'row'};
+  align-items: ${(props) => props.align || 'stretch'};
+  justify-content: ${(props) => props.justify || 'stretch'};
+  margin: ${({ margin }) => margin || '0'};
+  ${(props) => props.underline && css`
+    border-bottom: 1px solid ${props=>props.theme.colors.primary};
+  `}
+`;
